#896 move kernelschedule tests into separate file
